+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 ### Changed
 - Cleanup: removed unused `pixel_size` argument from `novae.spatial_neighbors`
 
+### Fixed
+- Upgrade sopa to import `sopa.constants` instead of `sopa._constants` (#38)
+
 ## [1.0.1] - 2025-12-10
 
 🎉 Novae is now [published in Nature Methods](https://www.nature.com/articles/s41592-025-02899-6)!